fix: rename module to include rural html
