+++ src/modules/CreateHtmlForRM/utils/brasaoLogo.ts
