+++ src/modules/CreateHtmlForRM/utils/renderImoveisRurais.ts
