+++ src/modules/CreateHtmlForRM/utils/verificarSeTodosSaoVazios.ts
